+++ src/main/kotlin/ru/spb/ivsamokhvalov/example/demo/camunda/CamundaConstants.kt
@@ -1,15 +1,14 @@
 package ru.spb.ivsamokhvalov.example.demo.camunda
 
 object CamundaConstants {
-
+    //posting
     const val POSTING_PROCESS_KEY = "CreatePostingProcessId"
-    const val ORDER_PROCESS_KEY = "CreateOrderProcessId"
+
+    //variables
     const val POSTING_ID = "postingId"
     const val POSTING_STATUS = "postingStatus"
     const val AWAITING_PAYMENT = "awaitingPayment"
     const val AWAITING_RECEIVED = "awaitingReceived"
-    fun buildPostingProcessBusinessKey(postingId: Long) = "posting_${postingId}"
-    fun buildOrderProcessBusinessKey(postingId: Long) = "order_${postingId}"
 
     //messageIds
     const val POSTING_IS_PAID = "postingIsPaid"
@@ -19,10 +18,19 @@ object CamundaConstants {
     const val POSTING_CANCELLED = "postingCancelled"
     const val ORDER_CANCELLED = "orderCancelled"
 
+    //order
+    const val ORDER_PROCESS_KEY = "CreateOrderProcessId"
 
+    //variables
     const val ORDER_ID = "orderId"
     const val CURRENT_ORDER_STATUS = "currentStatus"
     const val NEXT_ORDER_STATUS = "nextStatus"
+
+    //messageIds
     const val RECALCULATE_ORDER_STATUS = "recalculateOrderStatus"
+
+    fun buildPostingProcessBusinessKey(postingId: Long) = "posting_${postingId}"
+    fun buildOrderProcessBusinessKey(postingId: Long) = "order_${postingId}"
+
 
 }
