+++ src/main/kotlin/ru/spb/ivsamokhvalov/example/demo/camunda/DemoCamundaApplication.kt
@@ -4,9 +4,7 @@ import org.camunda.bpm.spring.boot.starter.annotation.EnableProcessApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 import org.springframework.cache.annotation.EnableCaching
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing
 
-@EnableJpaAuditing
 @SpringBootApplication
 @EnableCaching
 @EnableProcessApplication
