+++ src/main/kotlin/ru/spb/ivsamokhvalov/example/demo/camunda/configuations/Configurations.kt
@@ -11,28 +11,15 @@ import org.springframework.web.filter.CorsFilter
 @Configuration
 class Configurations {
 
-//    @Bean
-//    fun corsConfigurationSource(): CorsConfigurationSource {
-//        val configuration = CorsConfiguration()
-//        configuration.allowCredentials = true
-//        configuration.addAllowedOrigin("*")
-//        configuration.addAllowedHeader("*")
-//        configuration.addAllowedMethod("*")
-//        val source = UrlBasedCorsConfigurationSource()
-//        source.registerCorsConfiguration("/**", configuration)
-//        return source
-//    }
-
-
     @Bean
     fun processCorsFilter(): FilterRegistrationBean<*> {
         val source = UrlBasedCorsConfigurationSource()
-        val config = CorsConfiguration()
-        config.allowCredentials = true
-//        config.addAllowedOrigin("*")
-        config.allowedOriginPatterns = listOf("*")
-        config.addAllowedHeader("*")
-        config.addAllowedMethod("*")
+        val config = CorsConfiguration().also {
+            it.allowCredentials = true
+            it.allowedOriginPatterns = listOf("*")
+            it.addAllowedHeader("*")
+            it.addAllowedMethod("*")
+        }
         source.registerCorsConfiguration("/**", config)
         val bean = FilterRegistrationBean(CorsFilter(source))
         bean.order = 0
