--- src/main/resources/bpmn/FirstExample.bpmn
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<bpmn:definitions xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" xmlns:modeler="http://camunda.org/schema/modeler/1.0" id="Definitions_16vp84e" targetNamespace="http://bpmn.io/schema/bpmn" exporter="Camunda Modeler" exporterVersion="5.0.0" modeler:executionPlatform="Camunda Platform" modeler:executionPlatformVersion="7.17.0">
-  <bpmn:collaboration id="Collaboration_0514f79">
-    <bpmn:participant id="Participant_1qp1p8l" name="svszdgf" processRef="Process_03gfg8y" />
-  </bpmn:collaboration>
-  <bpmn:process id="Process_03gfg8y" name="sdfsf" isExecutable="false">
-    <bpmn:startEvent id="StartEvent_1">
-      <bpmn:outgoing>Flow_195z2uy</bpmn:outgoing>
-    </bpmn:startEvent>
-    <bpmn:endEvent id="Event_19fpibj">
-      <bpmn:incoming>Flow_195z2uy</bpmn:incoming>
-    </bpmn:endEvent>
-    <bpmn:sequenceFlow id="Flow_195z2uy" sourceRef="StartEvent_1" targetRef="Event_19fpibj" />
-  </bpmn:process>
-  <bpmn:message id="Message_0g3nma4" name="testMessage" />
-  <bpmndi:BPMNDiagram id="BPMNDiagram_1">
-    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="Collaboration_0514f79">
-      <bpmndi:BPMNShape id="Participant_1qp1p8l_di" bpmnElement="Participant_1qp1p8l" isHorizontal="true">
-        <dc:Bounds x="158" y="57" width="600" height="333" />
-        <bpmndi:BPMNLabel />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNEdge id="Flow_195z2uy_di" bpmnElement="Flow_195z2uy">
-        <di:waypoint x="248" y="180" />
-        <di:waypoint x="380" y="180" />
-        <di:waypoint x="380" y="117" />
-        <di:waypoint x="512" y="117" />
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNShape id="Event_0lpvfvl_di" bpmnElement="StartEvent_1">
-        <dc:Bounds x="212" y="162" width="36" height="36" />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="Event_19fpibj_di" bpmnElement="Event_19fpibj">
-        <dc:Bounds x="512" y="99" width="36" height="36" />
-      </bpmndi:BPMNShape>
-    </bpmndi:BPMNPlane>
-  </bpmndi:BPMNDiagram>
-</bpmn:definitions>
